+++ stac/src/catalog.rs
@@ -1,4 +1,4 @@
-use crate::{Href, Link, Links, STAC_VERSION};
+use crate::{Error, Extensions, Href, Link, Links, Result, STAC_VERSION};
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 
@@ -20,6 +20,10 @@ pub const CATALOG_TYPE: &str = "Catalog";
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub struct Catalog {
     /// Set to `"Catalog"` if this Catalog only implements the `Catalog` spec.
+    #[serde(
+        deserialize_with = "deserialize_type",
+        serialize_with = "serialize_type"
+    )]
     pub r#type: String,
 
     /// The STAC version the `Catalog` implements.
@@ -97,6 +101,37 @@ impl Links for Catalog {
     fn links_mut(&mut self) -> &mut Vec<Link> {
         &mut self.links
     }
+}
+
+impl TryFrom<Catalog> for Map<String, Value> {
+    type Error = Error;
+    fn try_from(catalog: Catalog) -> Result<Self> {
+        if let serde_json::Value::Object(object) = serde_json::to_value(catalog)? {
+            Ok(object)
+        } else {
+            panic!("all STAC catalogs should serialize to a serde_json::Value::Object")
+        }
+    }
+}
+
+impl Extensions for Catalog {
+    fn extensions(&self) -> Option<&[String]> {
+        self.extensions.as_deref()
+    }
+}
+
+fn deserialize_type<'de, D>(deserializer: D) -> std::result::Result<String, D::Error>
+where
+    D: serde::de::Deserializer<'de>,
+{
+    crate::deserialize_type(deserializer, CATALOG_TYPE)
+}
+
+fn serialize_type<S>(r#type: &String, serializer: S) -> std::result::Result<S::Ok, S::Error>
+where
+    S: serde::ser::Serializer,
+{
+    crate::serialize_type(r#type, serializer, CATALOG_TYPE)
 }
 
 #[cfg(test)]